Refactor
I apologize for the shoddy quality of the things that were checked in
previously. My mind wasn't totally on my work due to family issues.

For the last week, we'd been taking care of an injured kitty. On Friday,
she passed away.

It seems that when I'm overly stressed, I do things stupidly, so I fixed
the way some things were done:

The inputs are now string[] instead of char[ , ] for recognizing the OCR
digits

The string[] versions of the OCR characters have been moved into a
project called KataBankOCR.Shared in an effort to be DRY (since they are
used by both the test project and code project)

This is way more in the way of changes than I normally check in at one
time on a refactor, but to do it otherwise would have resulted in a
broken build.

No new functionality added as this was just a refactor commit.

All tests green.