Added reading input from file and bugfix
Added ability to read from a file.

It also seems that one of my methods was still just stubbed out, but
causes the tests to pass (due to not having enough/the right tests).

Added extra tests for triangulation and actually implemented the
function that I forgot to write like the dummy I am.

I must have been channeling my inner Ed Gruberman.
Forget a boot to the head. I needed a shoe store to the head lol