PredictTheNumber - fixed bug Unit test was flawed, allowing a pair of bugs to be missed. Recursive call for NumberOfTransformations used arguments in the wrong order. Fencepost error in NumberOfTransformations
Added integration tests for checking value of positions 0-7 (which caused the other bugs to be revealed).

Basically the tests were previously green due to a really weird fluke. However, everything up to this point seems to be working now.

All tests green (for what that matters).

TL;DR - I'm a dummy and it's amazing that anything ever works =]